+++ classes/CommunityPoll.js
@@ -1,6 +1,6 @@
 const { Collection } = require('discord.js');
 const EmbedBase = require('./components/EmbedBase');
-const Firebase = require('./FirebaseAPI');
+import * as Firebase from '../api';
 const XPService = require('./services/XPService');
 
 class CommunityPoll {
