+++ commands/admin/inspect.js
@@ -1,6 +1,6 @@
 const Command = require('../../classes/Command');
 const EmbedBase = require('../../classes/components/EmbedBase');
-const Firebase = require('../../classes/FirebaseAPI');
+import * as Firebase from '../../api';
 
 class inspect extends Command {
     constructor(bot) {
